pyproject.toml
[tool.poetry]
name = "apache-airflow-superset-poc"
version = "0.1.0"
description = ""
authors = ["Randy Topliffe <randytopliffe@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.8"
mimesis = "^11.1.0"

[tool.poetry.group.dev.dependencies]
apache-airflow = "^2.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
